+++ cmd/rqlited/main.go
@@ -303,7 +303,6 @@ func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
 
 	// Set optional parameters on store.
 	str.RaftLogLevel = cfg.RaftLogLevel
-	str.NoFreeListSync = cfg.RaftNoFreelistSync
 	str.ShutdownOnRemove = cfg.RaftShutdownOnRemove
 	str.SnapshotThreshold = cfg.RaftSnapThreshold
 	str.SnapshotInterval = cfg.RaftSnapInterval