+++ CHANGELOG.md
@@ -39,6 +39,7 @@ When officially released 8.0 will support (mostly) seamless upgrades from the 7.
 - [PR #1399](https://github.com/rqlite/rqlite/pull/1399): Better trailing flags error message.
 - [PR #1404](https://github.com/rqlite/rqlite/pull/1404): Add an interface between Store and Snapshot Store.
 - [PR #1410](https://github.com/rqlite/rqlite/pull/1410), [PR #1412](https://github.com/rqlite/rqlite/pull/1412): Implement simpler WAL-based snapshotting.
+- [PR #1413](https://github.com/rqlite/rqlite/pull/1413): Remove `-raft-no-freelist-sync` command line flag.
 
 ## 7.21.4 (July 8th 2023)
 ### Implementation changes and bug fixes